2.2 JPQL, Spring ORM, DAO на основе Spring ORM + JPA

+++ book-library/src/main/java/ru/diasoft/spring/booklibrary/service/BookService.java
@@ -9,7 +9,7 @@ public interface BookService {
 
     String getBookInfo(Long bookId);
 
-    int getNumberOfBooks();
+    long getNumberOfBooks();
 
     void deleteBook(Long bookId);
